abstract database
